+++ server/src/sql_tcp_backend_handler.rs
@@ -15,8 +15,8 @@ use std::collections::HashSet;
 use tracing::{debug, instrument};
 
 fn gen_random_string(len: usize) -> String {
-    use rand::{Rng, SeedableRng, distributions::Alphanumeric, rngs::SmallRng};
-    let mut rng = SmallRng::from_entropy();
+    use rand::{Rng, distributions::Alphanumeric, rngs::OsRng};
+    let mut rng = OsRng;
     std::iter::repeat(())
         .map(|()| rng.sample(Alphanumeric))
         .map(char::from)
@@ -136,6 +136,10 @@ impl TcpBackendHandler for SqlBackendHandler {
             .is_none()
         {
             debug!("User not found");
+            return Ok(None);
+        }
+        if self.is_user_disabled(user).await? {
+            debug!("Password reset denied for disabled user");
             return Ok(None);
         }